+++ .gitignore
@@ -18,3 +18,9 @@ text-ui-test/EXPECTED-UNIX.TXT
 
 # class files
 *.class
+
+# Ignore Gradle project-specific cache directory
+.gradle
+
+# Ignore Gradle build output directory
+build
