+++ plugin/src/main/java/org/example/IpPlugin.java
@@ -1,0 +1,19 @@
+/*
+ * This Java source file was generated by the Gradle 'init' task.
+ */
+package org.example;
+
+import org.gradle.api.Project;
+import org.gradle.api.Plugin;
+
+/**
+ * A simple 'hello world' plugin.
+ */
+public class IpPlugin implements Plugin<Project> {
+    public void apply(Project project) {
+        // Register a task
+        project.getTasks().register("greeting", task -> {
+            task.doLast(s -> System.out.println("Hello from plugin 'org.example.greeting'"));
+        });
+    }
+}
