--- src/main/java/CoatException.java
@@ -1,5 +1,0 @@
-public class CoatException extends RuntimeException {
-    public CoatException(String message) {
-        super(message);
-    }
-}
