--- src/main/java/Messages.java
@@ -1,36 +1,0 @@
-/**
- * This class contains ASCII art messages generated online.
- * Escape characters for backslashes have been added to the ASCII art.
- *
- * @author @AmanZainal
- */
-
-public class Messages {
-
-    /**
-     * The ASCII art representing a welcome message.
-     * Includes a stylized text and an ASCII art cat saying "meow."
-     */
-    public static final String WELCOME =
-            "  _____            _     _   _             _____      _   _           _   \n" +
-            " / ____|          | |   | | | |           / ____|    | | | |         | |  \n" +
-            "| |     ___   __ _| |_  | |_| |__   ___  | |     __ _| |_| |__   ___ | |_ \n" +
-            "| |    / _ \\ / _` | __| | __| '_ \\ / _ \\ | |    / _` | __| '_ \\ / _ \\| __|\n" +
-            "| |___| (_) | (_| | |_  | |_| | | |  __/ | |___| (_| | |_| |_) | (_) | |_ \n" +
-            "\\_____/\\___/ \\__,_|\\__|  \\__|_| |_|\\___| \\_____|\\__,_|\\__|_.__/ \\___/ \\__|\n" +
-            "                                                                          \n" +
-
-            "                        _    _      _ _       _                        \n" +
-            "                       | |  | |    | | |     | |         /\\_/\\  meow    \n" +
-            "  ___  __ _ _   _ ___  | |__| | ___| | | ___ | |        ( o.o )       \n" +
-            " / __|/ _` | | | / __| |  __  |/ _ \\ | |/ _ \\| |         > ^ <         \n" +
-            " \\__ \\ (_| | |_| \\__ \\ | |  | |  __/ | | (_) |_|                   \n" +
-            " |___/\\__,_|\\__, |___/ |_|  |_|\\___|_|_|\\___/(_)                   \n" +
-            "             __/ |                                                     \n" +
-            "            |___/                                                      \n" +
-            "\nWhat are we doing today? ~(^o.o^)\nSay 'help' if you want to know the tricks I can do!";
-
-    public static final String EXIT =
-    "Aw, see you soon! :3\n";
-
-}
