--- src/main/java/Task.java
@@ -1,30 +1,0 @@
-public class Task {
-    private String description;
-    private boolean isDone;
-
-    public Task(String description) {
-        this.description = description;
-        this.isDone = false;
-    }
-
-    public String getDescription() {
-        return description;
-    }
-
-    public boolean isDone() {
-        return isDone;
-    }
-
-    public void markAsDone() {
-        isDone = true;
-    }
-
-    public void markUndone() {
-        isDone = false;
-    }
-
-    @Override
-    public String toString() {
-        return String.format("[%s] %s", isDone ? "X" : " ", description);
-    }
-}
