--- src/main/java/UnknownCommand.java
@@ -1,11 +1,0 @@
-import java.util.List;
-
-public class UnknownCommand extends Command<List<Task>> {
-    public UnknownCommand() {
-        super("", List.of());
-    }
-
-    List<Task> execute(List<Task> tasks) {
-        return null;
-    }
-}
