+++ src/main/java/duke/command/UnknownCommand.java
@@ -1,0 +1,16 @@
+package duke.command;
+
+import duke.exception.*;
+import duke.task.*;
+import duke.ui.*;
+import java.util.List;
+
+public class UnknownCommand extends Command<List<Task>> {
+    public UnknownCommand() {
+        super("", List.of());
+    }
+
+    public List<Task> execute(List<Task> tasks) {
+        return null;
+    }
+}
