+++ src/main/java/duke/task/Deadline.java
@@ -1,0 +1,17 @@
+package duke.task;
+
+public class Deadline extends Task {
+    private String deadline;
+    private TaskType type;
+
+    public Deadline(String description, String deadline) {
+        super(description);
+        this.deadline = deadline;
+        this.type = TaskType.D;
+    }
+
+    @Override
+    public String toString() {
+        return String.format("[%s]%s (by: %s)", type, super.toString(), deadline);
+    }
+}
