+++ src/main/java/duke/task/TaskType.java
@@ -1,0 +1,7 @@
+package duke.task;
+
+public enum TaskType {
+    T,
+    D,
+    E
+}
