+++ src/main/java/duke/task/ToDo.java
@@ -1,0 +1,15 @@
+package duke.task;
+
+public class ToDo extends Task {
+    private TaskType type;
+
+    public ToDo(String description) {
+        super(description);
+        this.type = TaskType.T;
+    }
+
+    @Override
+    public String toString() {
+        return String.format("[%s]%s", type, super.toString());
+    }
+}
